Rename submodules under `cupyx.scipy.ndimage`
