+++ cupyx/scipy/ndimage/__init__.py
@@ -1,45 +1,45 @@
-from cupyx.scipy.ndimage.filters import correlate  # NOQA
-from cupyx.scipy.ndimage.filters import convolve  # NOQA
-from cupyx.scipy.ndimage.filters import correlate1d  # NOQA
-from cupyx.scipy.ndimage.filters import convolve1d  # NOQA
-from cupyx.scipy.ndimage.filters import uniform_filter1d  # NOQA
-from cupyx.scipy.ndimage.filters import uniform_filter  # NOQA
-from cupyx.scipy.ndimage.filters import gaussian_filter1d  # NOQA
-from cupyx.scipy.ndimage.filters import gaussian_filter  # NOQA
-from cupyx.scipy.ndimage.filters import prewitt  # NOQA
-from cupyx.scipy.ndimage.filters import sobel  # NOQA
-from cupyx.scipy.ndimage.filters import generic_laplace  # NOQA
-from cupyx.scipy.ndimage.filters import laplace  # NOQA
-from cupyx.scipy.ndimage.filters import gaussian_laplace  # NOQA
-from cupyx.scipy.ndimage.filters import generic_gradient_magnitude  # NOQA
-from cupyx.scipy.ndimage.filters import gaussian_gradient_magnitude  # NOQA
-from cupyx.scipy.ndimage.filters import minimum_filter  # NOQA
-from cupyx.scipy.ndimage.filters import maximum_filter  # NOQA
-from cupyx.scipy.ndimage.filters import minimum_filter1d  # NOQA
-from cupyx.scipy.ndimage.filters import maximum_filter1d  # NOQA
-from cupyx.scipy.ndimage.filters import median_filter  # NOQA
-from cupyx.scipy.ndimage.filters import rank_filter  # NOQA
-from cupyx.scipy.ndimage.filters import percentile_filter  # NOQA
-from cupyx.scipy.ndimage.filters import generic_filter  # NOQA
-from cupyx.scipy.ndimage.filters import generic_filter1d  # NOQA
+from cupyx.scipy.ndimage._filters import correlate  # NOQA
+from cupyx.scipy.ndimage._filters import convolve  # NOQA
+from cupyx.scipy.ndimage._filters import correlate1d  # NOQA
+from cupyx.scipy.ndimage._filters import convolve1d  # NOQA
+from cupyx.scipy.ndimage._filters import uniform_filter1d  # NOQA
+from cupyx.scipy.ndimage._filters import uniform_filter  # NOQA
+from cupyx.scipy.ndimage._filters import gaussian_filter1d  # NOQA
+from cupyx.scipy.ndimage._filters import gaussian_filter  # NOQA
+from cupyx.scipy.ndimage._filters import prewitt  # NOQA
+from cupyx.scipy.ndimage._filters import sobel  # NOQA
+from cupyx.scipy.ndimage._filters import generic_laplace  # NOQA
+from cupyx.scipy.ndimage._filters import laplace  # NOQA
+from cupyx.scipy.ndimage._filters import gaussian_laplace  # NOQA
+from cupyx.scipy.ndimage._filters import generic_gradient_magnitude  # NOQA
+from cupyx.scipy.ndimage._filters import gaussian_gradient_magnitude  # NOQA
+from cupyx.scipy.ndimage._filters import minimum_filter  # NOQA
+from cupyx.scipy.ndimage._filters import maximum_filter  # NOQA
+from cupyx.scipy.ndimage._filters import minimum_filter1d  # NOQA
+from cupyx.scipy.ndimage._filters import maximum_filter1d  # NOQA
+from cupyx.scipy.ndimage._filters import median_filter  # NOQA
+from cupyx.scipy.ndimage._filters import rank_filter  # NOQA
+from cupyx.scipy.ndimage._filters import percentile_filter  # NOQA
+from cupyx.scipy.ndimage._filters import generic_filter  # NOQA
+from cupyx.scipy.ndimage._filters import generic_filter1d  # NOQA
 
-from cupyx.scipy.ndimage.fourier import fourier_gaussian  # NOQA
-from cupyx.scipy.ndimage.fourier import fourier_shift  # NOQA
-from cupyx.scipy.ndimage.fourier import fourier_uniform  # NOQA
+from cupyx.scipy.ndimage._fourier import fourier_gaussian  # NOQA
+from cupyx.scipy.ndimage._fourier import fourier_shift  # NOQA
+from cupyx.scipy.ndimage._fourier import fourier_uniform  # NOQA
 
-from cupyx.scipy.ndimage.interpolation import affine_transform  # NOQA
-from cupyx.scipy.ndimage.interpolation import map_coordinates  # NOQA
-from cupyx.scipy.ndimage.interpolation import rotate  # NOQA
-from cupyx.scipy.ndimage.interpolation import shift  # NOQA
-from cupyx.scipy.ndimage.interpolation import zoom  # NOQA
+from cupyx.scipy.ndimage._interpolation import affine_transform  # NOQA
+from cupyx.scipy.ndimage._interpolation import map_coordinates  # NOQA
+from cupyx.scipy.ndimage._interpolation import rotate  # NOQA
+from cupyx.scipy.ndimage._interpolation import shift  # NOQA
+from cupyx.scipy.ndimage._interpolation import zoom  # NOQA
 
-from cupyx.scipy.ndimage.measurements import label  # NOQA
-from cupyx.scipy.ndimage.measurements import sum  # NOQA
-from cupyx.scipy.ndimage.measurements import mean  # NOQA
-from cupyx.scipy.ndimage.measurements import variance  # NOQA
-from cupyx.scipy.ndimage.measurements import standard_deviation  # NOQA
+from cupyx.scipy.ndimage._measurements import label  # NOQA
+from cupyx.scipy.ndimage._measurements import sum  # NOQA
+from cupyx.scipy.ndimage._measurements import mean  # NOQA
+from cupyx.scipy.ndimage._measurements import variance  # NOQA
+from cupyx.scipy.ndimage._measurements import standard_deviation  # NOQA
 
-from cupyx.scipy.ndimage.morphology import grey_erosion  # NOQA
-from cupyx.scipy.ndimage.morphology import grey_dilation  # NOQA
-from cupyx.scipy.ndimage.morphology import grey_closing  # NOQA
-from cupyx.scipy.ndimage.morphology import grey_opening  # NOQA
+from cupyx.scipy.ndimage._morphology import grey_erosion  # NOQA
+from cupyx.scipy.ndimage._morphology import grey_dilation  # NOQA
+from cupyx.scipy.ndimage._morphology import grey_closing  # NOQA
+from cupyx.scipy.ndimage._morphology import grey_opening  # NOQA
