+++ cupyx/scipy/ndimage/_filters.py
