+++ cupyx/scipy/ndimage/_fourier.py
