+++ cupyx/scipy/ndimage/_interpolation.py
