+++ cupyx/scipy/signal/bsplines.py
@@ -39,5 +39,5 @@ def sepfir2d(input, hrow, hcol):
     hrow = hrow.astype(dtype, copy=False)
     hcol = hcol.astype(dtype, copy=False)
     filters = (hcol[::-1], hrow[::-1])
-    return cupyx.scipy.ndimage.filters._run_1d_correlates(
+    return cupyx.scipy.ndimage._filters._run_1d_correlates(
         input, (0, 1), lambda i: filters[i], None, 'reflect', 0)
